+++ Projet Piscine/Projet Piscine/Sommet.h
@@ -33,7 +33,7 @@ public:
     void calc_icd(const size_t&ordre);
     void calc_vp(std::map<Sommet*,double>&somme);
     void indice_vp(std::map<Sommet*,double>&somme,const double& lambda);
-    void calc_ici_naif(int total,int a);
+    void calc_ici_naif(double total,double a);
     int k_connexe(const Sommet* arrive)const;
     void k_connexe(int& nombre_chemin,std::map<const Arete*,bool>& arete,std::map<const Sommet*,bool>&sommet,const Sommet*arrive)const;
